+++ client/src/components/Skills/display/SkillDetail.tsx
@@ -5,6 +5,7 @@ import { TooltipAnchor } from '@librechat/client';
 import type { TSkill } from 'librechat-data-provider';
 import { useLocalize, useAuthContext, useSkillPermissions } from '~/hooks';
 import SkillMarkdownRenderer from './SkillMarkdownRenderer';
+import { parseFrontmatter } from '../utils';
 import DeleteSkill from '../dialogs/DeleteSkill';
 import { ShareSkill } from '../buttons';
 import { cn } from '~/utils';
@@ -15,60 +16,7 @@ interface SkillDetailProps {
   onDelete?: () => void;
 }
 
-/**
- * Strip YAML frontmatter (`---\n...\n---`) from a SKILL.md body and return
- * the frontmatter fields as a key-value map + the remaining body.
- */
-function parseFrontmatter(raw: string): {
-  fields: Array<{ key: string; value: string }>;
-  body: string;
-} {
-  const trimmed = raw.trim();
-  if (!trimmed.startsWith('---')) {
-    return { fields: [], body: raw };
-  }
-  const after = trimmed.slice(3);
-  const closingIdx = after.indexOf('\n---');
-  if (closingIdx === -1) {
-    return { fields: [], body: raw };
-  }
-
-  const block = after.slice(0, closingIdx);
-  const body = after.slice(closingIdx + 4).trim();
-
-  const fields: Array<{ key: string; value: string }> = [];
-  const lines = block.split('\n');
-  for (let i = 0; i < lines.length; i++) {
-    const line = lines[i];
-    const colon = line.indexOf(':');
-    if (colon === -1) {
-      continue;
-    }
-    const key = line.slice(0, colon).trim();
-    let value = line.slice(colon + 1).trim();
-    if (key.toLowerCase() === 'name' || key.toLowerCase() === 'description') {
-      continue;
-    }
-    if (!value) {
-      const items: string[] = [];
-      while (i + 1 < lines.length) {
-        const next = lines[i + 1];
-        const item = next.trim();
-        if (!item.startsWith('-')) {
-          break;
-        }
-        items.push(item.slice(1).trim());
-        i++;
-      }
-      value = items.join(',');
-    }
-    if (key && value) {
-      fields.push({ key, value });
-    }
-  }
-
-  return { fields, body };
-}
+const SKIP_KEYS = new Set(['name', 'description']);
 
 function ViewToggle({
   viewMode,
@@ -130,7 +78,7 @@ export default function SkillDetail({ skill, onEdit, onDelete }: SkillDetailProp
     : undefined;
 
   const { fields: frontmatterFields, body: cleanBody } = useMemo(
-    () => parseFrontmatter(skill.body ?? ''),
+    () => parseFrontmatter(skill.body ?? '', SKIP_KEYS),
     [skill.body],
   );
