+++ client/src/components/Skills/utils/index.ts
@@ -1,2 +1,4 @@
+export { parseFrontmatter } from './frontmatter';
 export { parseSkillMd } from './parseSkillMd';
+export type { FrontmatterField, ParsedFrontmatter } from './frontmatter';
 export type { ParsedSkillMd } from './parseSkillMd';
